\docType{class}
\name{athlete-class}
\alias{athlete-class}
\title{An S4 class to represent a athlete.}
\description{
  An S4 class to represent a athlete.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the athlete}

\item{\code{sex}}{sex of the athlete}

\item{\code{HRest}}{resting pulse of the athlete}

\item{\code{HRMax}}{the maximum heart rate of the athlete}

\item{\code{Zone1}}{50\%-60\% of maximum heart rate}

\item{\code{Zone2}}{60\%-70\% of maximum heart rate}

\item{\code{Zone3}}{70\%-80\% of maximum heart rate}

\item{\code{Zone4}}{80\%-90\% of maximum heart rate}

\item{\code{Zone5}}{90\%-100\% of maximum heart rate}
}}
