\docType{methods}
\name{plot_route}
\alias{plot_route}
\alias{plot_route,training-method}
\title{Plot a single training session}
\usage{
  plot_route(obj, num = 1)

\S4method{plot_route}{training}(obj, num = 1)
}
\arguments{
  \item{obj}{
    an object of class training
  }
  \item{num}{
    number of trainig session. default value: 1
  }
}
\description{
  Get an object of class training and plot one training session. Click on the Route to get further information, e.g. date, duration, distance, average speed and altitude.
}
