\docType{methods}
\name{plot_trimp_zones}
\alias{plot_trimp_zones}
\alias{plot_trimp_zones,training-method}
\title{Plot zonal trimp}
\usage{
  plot_trimp_zones(obj, num = 1)

\S4method{plot_trimp_zones}{training}(obj, num = 1)
}
\arguments{
  \item{obj}{
    an object of class training
  }
  \item{num}{
    number of trainig session. default value: 1
  }
}
\description{
  Get an object of class training and plot the dwell time in minutes in a zone.
}


