\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,activity-method}
\alias{summary,training-method}
\title{gives a summary of running data}
\usage{
  summary(obj)

\S4method{summary}{activity}(obj)
\S4method{summary}{training}(obj)
}
\arguments{
  \item{obj}{an object of class activity or training}
}
\description{
  This method gives a detailed summary of the trainings activity. It prints information about the distance, duration, heart rate, climb and pace.
}
