\docType{class}
\name{training-class}
\alias{training-class}
\title{An S4 class to represent an athlete with his runs.}
\description{
  An S4 class to represent an athlete with his runs.
}
\section{Slots}{

\describe{
\item{\code{activity}}{list of objects from class activity}

\item{\code{athlete}}{object of class athlete}
}}
