\name{training}
\alias{training}
\title{
  creates an object of class training
}
\description{
  Get a list of activity objects and a object of class athlete. It creates a new object of class training
}
\usage{
  training(actls, ath)
}
\arguments{
  \item{actls}{
    list with activity class objects
  }
  \item{ath}{
    object of class athlete
  }
}
\value{
  \item{tr}{
    the new object of class training
  }
}
