\name{trimp_print_fct}
\alias{trimp_print_fct}
\title{
  prints a summary of Trimp's
}
\description{
  Get an object of class training and the Trimp's of that trainings data. Prints an overview of the training and the Trimp's.
}
\usage{
  trimp_print_fct(obj, trimp)
}
\arguments{
  \item{obj}{
    object of class training
  }
  \item{trimp}{
    Trimp values of the appropriate trainings data
  }
}
