\docType{methods}
\name{trimp_z_fct}
\alias{trimp_z_fct}
\alias{trimp_z_fct,training-method}
\title{
  Calculates the zonal Trimp
}
\usage{
  trimp_z_fct(obj)

\S4method{trimp_z_fct}{training}(obj)
}
\arguments{
  \item{obj}{
    object of class training
  }
}
\value{
  \item{trimp_zo}{
    all zonal Trimp's of the training activity
  }
}
\description{
  Get an object of class training and calculates to every training activity the zonal Trimp
}
