\docType{methods}
\name{trimp_zone_values}
\alias{trimp_zone_values}
\alias{zone_values}
\title{
  Calculates the zone for each sequence
}
\usage{
  trimp_zone_values(obj)

\S4method{trimp_zone_values}{training}(obj)
}
\arguments{
  \item{obj}{
    an object of class training
  }
}
\description{
  Calculates the trimp zone for each sequence in all given activity objects.
}

